refactor: add context and provider in layout
