+++ src/context/theme-context.ts
@@ -2,6 +2,6 @@ import { createContext } from 'react';
 import { ThemeContextType } from '@/types/types';
 
 export const ThemeContext = createContext<ThemeContextType>({
-  theme: 'light',
-  setTheme: () => {},
+  themeValue: 'light',
+  setThemeValue: () => {},
 });
