+++ src/shared/config/theme.ts
@@ -1,14 +1,14 @@
 export const lightTheme = {
   token: {
-    colorPrimary: '#f59e0b',
-    colorBgBase: '#27272a',
+    colorPrimary: 'rgba(226, 186, 98, 1)',
+    colorBgBase: '#979797ff',
     fontSize: 16,
   },
   components: {
     Layout: {
-      headerBg: '#18181b',
-      footerBg: '#18181b',
-      bodyBg: '#27272a',
+      headerBg: '#4f4f52ff',
+      footerBg: '#4f4f52ff',
+      bodyBg: '#979797ff',
     },
   },
 };
@@ -21,7 +21,7 @@ export const darkTheme = {
   },
   components: {
     Layout: {
-      headerBg: '#f59e0b',
+      headerBg: '#18181b',
       footerBg: '#18181b',
       bodyBg: '#27272a',
     },
